% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliblist_advise_lipidomics.R
\name{caliblist_advise_lipidomics}
\alias{caliblist_advise_lipidomics}
\title{caliblist_advise_lipidomics}
\usage{
caliblist_advise_lipidomics(
  out,
  info,
  calibration_path,
  calibration_targetfile
)
}
\arguments{
\item{out}{List. It is the result from the \code{filter_advise_lipidomics} function.}

\item{info}{a string that tells if it's the deuterated or nonlabeled calibration}

\item{calibration_path}{Character string. It should be the folder in which calibration files
are located, already stored in the list 'out' (as out$folders$calibration_path).}

\item{calibration_targetfile}{Character string. It is the calibration target file.}
}
\value{
A calibration matrix
}
\description{
\code{caliblist_advise_lipidomics} is the function for creation of the calibration
data (concentration and area).
}
\details{
The first part of the calibration step creates a list of calibration information
from the calibration target files. This first part should be applied both to deuterated
files and to non-labeled files. The creation of the final calibration matrix is performed
outside the function. The calibration matrix has the correspondence between concentration
and area for each internal standard lipid species.
}
\note{
Last change 20/01/2022
}
