% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibplot_advise_lipidomics.R
\name{calibplot_advise_lipidomics}
\alias{calibplot_advise_lipidomics}
\title{calibplot_advise_lipidomics}
\usage{
calibplot_advise_lipidomics(
  out,
  cal_plot_path = "CalibrationPlot",
  cal_mat,
  intercept_flag = TRUE,
  lm_robust_flag = FALSE,
  lin_calibration = TRUE,
  plot_calibration = TRUE
)
}
\arguments{
\item{out}{List. It is the result from the \code{filter_advise_lipidomics} function.}

\item{cal_plot_path}{Character string. It is the name of folder in which the plots of the
calibration curves will be stored. Default = "CalibrationPlot".}

\item{cal_mat}{Dataframe. It is the calibration matrix, generated after the calibration step.}

\item{intercept_flag}{Logical value. If TRUE, linear model will be not forced to have the
intercept at zero. Default = FALSE.}

\item{lm_robust_flag}{Logical value. If TRUE, linear model will be performed in robust method
(by using the function by MASS package). Default = FALSE.}

\item{lin_calibration}{Logical value. If TRUE, a filter on concentration range of linearity is
applied, considering the min and max values in the internal standard file. Default  = TRUE.}
}
\value{
res: a list with results from filtering step, updated with the calibration matrix
and the linear model coefficients.
}
\description{
\code{calibplot_advise_lipidomics} is the function for creating the calibration
curves by the calibration matrix, and plotting the curves for the all the internal standard lipid species.
}
\details{
The second part of the calibration step calculates the coefficient values
(intercept and slope) from the application of a linear regression on the values of
the calibration matrix. The linear regression can be normal or robust, furthermore
the intercept can be force to be zero. All the curves (lines) can be plotted and
stored in the suitable folder.
}
\note{
Last change 02/11/2021
}
