% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lipidversion.R
\name{check_lipidversion}
\alias{check_lipidversion}
\title{check_lipidversion}
\usage{
check_lipidversion(
  file_paths,
  colnames_4.2 = c("LipidIon", "Class", "FattyAcid", "Ion", "ObsMz", "TopRT", "Area"),
  colnames_5.0 = c("LipidID", "ClassKey", "Adduct", "ObsMz", "TopRT", "Area")
)
}
\arguments{
\item{file_paths}{vector containing all the path files to be checked.}

\item{colnames_4.2}{Colnames used in the 4.2.29 LipidSearch version. By default c("LipidIon","Class","FattyAcid","Ion","ObsMz","TopRT","Area").}

\item{colnames_5.0}{Colnames used in the 4.2.29 LipidSearch version. By default c("LipidID", "ClassKey", "Adduct", "ObsMz" ,"TopRT","Area").}
}
\description{
\code{check_lipidversion} Run a check on LipidSearch version.
}
\details{
Run a check on LipidSearch version. If there is a problem, returns NULL. LipidSearch older than 4.2.29 are not supported.
Multiple LipidSearch version returns NULL.
}
\note{
Last change 05/07/2022
}
