% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_advise_lipidomics.R
\name{contrast_advise_lipidomics}
\alias{contrast_advise_lipidomics}
\title{Create contrast list}
\usage{
contrast_advise_lipidomics(
  out,
  design_vars = "",
  batch_vars = "",
  batch_type = "fit",
  batch_method = "given",
  ref = ""
)
}
\arguments{
\item{out}{The SummarizedExperiment result.}

\item{design_vars}{Character string. A design variable.}

\item{batch_vars}{Character. Batch effect variable.}

\item{batch_type}{Character. "remove" or "fit"}

\item{batch_method}{Method for the batch variable. "given" or....}
}
\description{
This function make a contrast list with maximum 2 variables.
}
