% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beautiful_radarchart.R
\name{create_beautiful_radarchart}
\alias{create_beautiful_radarchart}
\title{Create a customized spider plot}
\usage{
create_beautiful_radarchart(
  data,
  color = "#00AFBB",
  vlabels = colnames(data),
  caxislabels = NULL,
  title = NULL
)
}
\arguments{
\item{data}{A dataframe with 3 or 4 rows and some column (columns will be the variables in the spider plot). 
The rows must contain a row with maximum values and a row with mininum values, plus one or two rows with a "Codice_azienda".}

\item{color}{Function that provides colors to the plot. For example color = grDevices::hcl.colors(2, palette = "Dynamic").}

\item{vlabels}{Character vector for the names for variables.}

\item{caxislabels}{Range of values inside the spider plot (eg. caxislabels = c(0, 2, 5, 7, 10)).}

\item{title}{Character vector for the radarchart title.}
}
\description{
This function make a customized spider plot using the {fmsb} package.
}
