% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_advise_lipidomics.R
\name{enrichment_advise_lipidomics}
\alias{enrichment_advise_lipidomics}
\title{expdesign_advise_lipidomics}
\usage{
enrichment_advise_lipidomics(
  out,
  rank_c = "logFC",
  min_elem = 2,
  thresh = 0.05,
  k = ""
)
}
\arguments{
\item{out}{The result limma_result.}

\item{rank_c}{Character string. Value to plot. Possible options are: "logFC" , "P.Value", "adj.P.Val", "B"}

\item{min_elem}{Numeric. Internal value.}

\item{thresh}{p-value threshold. Possible options are 0.001, 0.01, 0.05.}

\item{k}{The name of the contrast.}

\item{sumexp}{List. The SummarizedExperiment result.}
}
\value{
write something
}
\description{
\code{enrichment_advise_lipidomics} is the function for enrichment analysis.
}
\details{
WRITE SOMETHING
}
\note{
Last change 17/11/2021
}
