% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdesign_advise_lipidomics.R
\name{expdesign_advise_lipidomics}
\alias{expdesign_advise_lipidomics}
\title{expdesign_advise_lipidomics}
\usage{
expdesign_advise_lipidomics(
  out,
  design_vars = "",
  batch_vars = "",
  batch_type = "",
  batch_method = "",
  rep_mean = FALSE,
  file_contrast = "contrast_list_mod.txt"
)
}
\arguments{
\item{out}{List. It is the result from the \code{sumexp_advise_lipidomics} function.}

\item{design_vars}{Character vector, with the names of the variables from the target file,
considered as conditions of comparison, related to replicates/samples. Default = "Product".}

\item{batch_vars}{Character vector, with the names of the variables from the target file,
considered as batch effects for the entire experiment. Default = "".}

\item{batch_type}{Character string. Different methodologies to cope with the eventual presence
of batch effects: "none", "remove", "fit". Default = "fit".}

\item{batch_method}{Character string. Different methodologies to cope with the selection of
batch_type = "fit". It can be "given" or "estimate". Default = "given".}

\item{rep_mean}{Logical value. It is for the selection of the dataset with the replicates
(FALSE) or the dataset with the samples (TRUE). Default = FALSE.}

\item{file_contrast}{Character string. It is the name of the file with the list of the
comparisons (contrasts) of interest. Default = "contrast_list.txt".}
}
\value{
res: a list with results from the Summarized Experiment step, updated with the
design matrix, the contrast matrix and the contrast vector (comparisons).
}
\description{
\code{expdesign_advise_lipidomics} is the function
}
\details{
This function generates the experimental design for the subsequent differential
analysis. At first, it provides the linear model formula with the selected design variables,
considering the eventual presence of batch effects. Then, after the generation of the
experimental design, it reads a list of comparisons of interest (contrast list) and generates
the contrast matrix for the next steps. In case of batch_type == "fit" and batch_method == "estimate",
the selected methodology creates surrogate variables to be added to the design matrix.
Further information on removing batch effects is in 'diffexp_advise_lipidomics" function.
}
\note{
Last change 17/12/2021
}
