% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_advise_lipidomics.R
\name{filter_advise_lipidomics}
\alias{filter_advise_lipidomics}
\title{filter_advise_lipidomics}
\usage{
filter_advise_lipidomics(
  out,
  ca_bound = c(14, 24),
  db_bound = c(0, 6),
  data_type = "LipidSearch"
)
}
\arguments{
\item{out}{List. It is the result from the \code{read_advise_lipidomics} function.}

\item{ca_bound}{Numerical vector. It is the range of the carbon atoms for each lipid
species: first number is the lower bound, second number is the upper bound. Default = c(14,24).}

\item{db_bound}{Numerical vector. It is the range of the double bonds for each lipid
species: first number is the lower bound, second number is the upper bound. Default = c(0,6).}

\item{data_type}{Character. Where come from the data. Can be "LipidSearch" or "LIQUID".}
}
\value{
res: a list with results from the reading step, updated with the filtered data, divided
into non-labeled data (lipid_filtered) and deuterated data (lipid_deuterated).
}
\description{
\code{filter_advise_lipidomics} is the function for filtering lipid data on
several features.
}
\details{
The filter step is composed of different filters applied on non-labeled files:
a) retention time, the retention times of the lipid species should be in the range reported
in the internal standard file for each class;
b) carbon atom, the carbon atoms of the lipid species should be in the range reported
in the internal standard file for each class, and should be an even number;
c) double bond, the double bonds of the lipid species should be in the range reported
in the internal standard file for each class;
d) duplicated area, the duplicated areas for a lipid species should be ingnored, conserving
only the minimum area for one lipid species (duplication is checkd on m/z values).
After this step, filtered data are stored separately from raw data.
}
\note{
Last change 17/12/2021
}
