% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sumexp.R
\name{filter_sumexp}
\alias{filter_sumexp}
\title{filter_sumexp}
\usage{
filter_sumexp(data, filt_targ = NULL, value_targ = NULL, filt_class = NULL)
}
\arguments{
\item{data}{A list containing both the SumExp objects and other information like replicates and data_type.}

\item{filt_targ}{Character. A column from the colData where to apply the filter (e.g. "Product").}

\item{value_targ}{Character. The values to be filtered on the colData (e.g. c("CTRL", "BC")).}

\item{filt_class}{Character. The lipid classes to be filtered (e.g. c("Cer", "PC", "PE")).}
}
\description{
\code{filter_sumexp} is the function for the filtering of the
SummarizedExperiment object.
}
\note{
Last change 23/06/2022
}
