% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linebreaks.R
\name{linebreaks}
\alias{linebreaks}
\title{filter_advise_lipidomics}
\usage{
linebreaks(n)
}
\arguments{
\item{n}{number of breaks}
}
\description{
\code{linebreaks} draw multiple break line (br())
}
\examples{
\dontrun{
linebreaks(10)   #add 10 breaks
}
}
