% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_advise_lipidomics.R
\name{read_advise_lipidomics}
\alias{read_advise_lipidomics}
\title{read_advise_lipidomics}
\usage{
read_advise_lipidomics(out, datapath, data_type = "LipidSearch", target_file)
}
\arguments{
\item{out}{List. It is the result from the \code{init_advise_lipidomics} function.}

\item{data_type}{Character string. The data type to read. Can be "LipidSearch" or "Liquid".
By default "LipidSearch".}

\item{target_file}{Dataframe. It should be the target file already stored in the list 'out'
(as out$targets$targetfile_lipidomics).}
}
\value{
res: a list with result from the initialization step, updated with the data to be
analyzed with the forthcoming steps.
}
\description{
\code{read_advise_lipidomics} is the function for reading and storing data
produced by LipidSearch, and to be analyzed with ADViSELipidomics pipeline.
}
\details{
The reading step reads and stores samples data (created by LipidSearch software),
checking if all the files listed in the target file are present in the suitable folder.
}
\note{
Last change 17/12/2021
}
