% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recovery_advise_lipidomics.R
\name{recovery_advise_lipidomics}
\alias{recovery_advise_lipidomics}
\title{recovery_advise_lipidomics}
\usage{
recovery_advise_lipidomics(out, intercept_flag)
}
\arguments{
\item{out}{List. It is the result from the \code{calibplot_advise_lipidomics} function.}
}
\value{
res: a list with results from calibration steps, updated with the recovery percentages
and the concentration matrix (feature matrix).
}
\description{
\code{recovery_advise_lipidomics} is the function for the computation of the
recovery percentages, and their application to define the concentration matrix (i.e. feature
matrix)
}
\details{
The recovery step is the computation of the recovery percentage for each internal
standard lipid species, and the application of the recovery percentage on all the filter
lipid species (non-labeled species). The complete algorithm of the computation and the
application of the recovery percentage is composed of two main steps:
a) the recovery percentage is calculated by the formula (q intercept, m slope,
NominalStdConcentration is a value of concentration from the internal standard file):
   - (Area/m)*100/NominalStdConcentration for intercept as zero;
   - (Area-q)/m)*100/NominalStdConcentration for intercept as not zero;
b) the recovery percentage is applied to the lipid species by the formula (ConcExpPerc
is the recovery percentage):
   - (Area/m)*100.00/ConcExpPerc for intercept as zero;
   - (Area-q)/m)*100.00/ConcExpPerc for intercept as not zero.
}
\note{
Last change 14/01/2022
}
