% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumexp_advise_lipidomics.R
\name{sumexp_advise_lipidomics}
\alias{sumexp_advise_lipidomics}
\title{sumexp_advise_lipidomics}
\usage{
sumexp_advise_lipidomics(out)
}
\arguments{
\item{out}{List. It is the result from the \code{recovery_advise_lipidomics} function.}
}
\value{
res: a list with results from the recovery step, updated with two SummerizedExperiment
objects with concentration matrix, target file information, lipid features and metadata.
}
\description{
\code{sumexp_advise_lipidomics} is the function for the creation of the
SummarizedExperiment object from concentration matrix, target file information, and
metadata (already stored in 'out' list)
}
\details{
The conversion step creates the SummarizeExperiment object from the concentration
matrix (i.e. the feature matrix), the target file data, and laboratory metadata. Moreover,
lipid features are parsed from the name of the lipid species. Two SummarizedExperiment objects
are stored: one with each single replicate, one with replicates averaged per sample.
}
\note{
Last change 17/12/2021
}
