% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_poor_mans_flatly2.R
\docType{data}
\name{theme_ADViSE}
\alias{theme_ADViSE}
\title{theme_ADViSE}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
theme_ADViSE
}
\value{
Object produced by shinyDashboardThemeDIY
}
\description{
Poor Man's Flatly theme for a shinydashboard application edited
}
\seealso{
\code{\link{shinyDashboardThemeDIY}}
}
\keyword{datasets}
